requirements.txt
Django~=4.0.4
djangorestframework~=3.13.1
django-ckeditor~=6.4.2
